- Fixed a few bugs in the explosion loader which prevented that   loader from working correctly. - Fixed a reference to zlib-win32-bin in the documentation. - Added full documentation for all the existing mesh object plugins   in CS. - Cleaned up mapfmt.txi documentation. i.e. removed all the   documentation related to sprite and particle systems. This   is now documented with the mesh object plugin documentation.   Also added documentation about MESHOBJ keyword.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@5750 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/loader/explo/exploldr.cpp
@@ -295,29 +295,29 @@ iBase* csExplosionLoader::Parse (const char* string, iEngine* engine)
         break;
       case CS_TOKEN_PARTRADIUS:
         {
-          int f;
-          ScanStr (params, "%d", &f);
+          float f;
+          ScanStr (params, "%f", &f);
           explostate->SetPartRadius (f);
         }
         break;
       case CS_TOKEN_SPREADPOS:
         {
-          int f;
-          ScanStr (params, "%d", &f);
+          float f;
+          ScanStr (params, "%f", &f);
           explostate->SetSpreadPos (f);
         }
         break;
       case CS_TOKEN_SPREADSPEED:
         {
-          int f;
-          ScanStr (params, "%d", &f);
+          float f;
+          ScanStr (params, "%f", &f);
           explostate->SetSpreadSpeed (f);
         }
         break;
       case CS_TOKEN_SPREADACCEL:
         {
-          int f;
-          ScanStr (params, "%d", &f);
+          float f;
+          ScanStr (params, "%f", &f);
           explostate->SetSpreadAcceleration (f);
         }
         break;
